requirements.txt
requests>=2.28
python-dotenv>=1.0
python-dateutil>=2.8
beautifulsoup4>=4.12
lxml>=4.9
